refactor: subroutine dec
